[MFS][TXN] - txn v2 skeleton.

+++ src/main/java/com/anika/mobilefinancialservice/repositories/TxnLogRepository.java
@@ -4,7 +4,7 @@ import com.anika.mobilefinancialservice.entity.TxnLogEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface TxnLogRepository extends JpaRepository<TxnLogEntity, Integer> {
+public interface TxnLogRepository extends JpaRepository<TxnLogEntity, Long> {
 
 
 }
